Moved utility functions useful in testing to _utils/testing.py from utils.py
 - Replicating the organization in Nilearn.
 - Functions moved:
   - write_fake_fmri_data()
   - generate_fake_fmri_data()
   - _write_fake_bold_img()
   - _basic_paradigm()
   - _basic_confounds()
   - create_fake_bids_dataset()
